# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TutorMatch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  student: "student",
  tutor: "tutor",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  tutor: "tutor",
  admin: "admin",
} as const;

export interface User {
  userId: number;
  name: string;
  email: string;
  role: UserRole;
  rating?: number | null;
  suspended: boolean;
  createdDate: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserRequest {
  name?: string;
}

export interface TutorProfile {
  tutorId: number;
  subjects: string[];
  education?: string | null;
  experience?: string | null;
  hourlyRate?: number | null;
  bio?: string | null;
  user?: User;
}

export interface UpdateTutorProfileRequest {
  subjects?: string[];
  education?: string;
  experience?: string;
  hourlyRate?: number;
  bio?: string;
}

export type QuestionStatus =
  (typeof QuestionStatus)[keyof typeof QuestionStatus];

export const QuestionStatus = {
  Open: "Open",
  Matched: "Matched",
  Scheduled: "Scheduled",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface Question {
  questionId: number;
  studentId: number;
  title: string;
  description: string;
  subject: string;
  attachmentUrl?: string | null;
  /** Duration in minutes */
  preferredDuration: number;
  optionalBudget?: number | null;
  status: QuestionStatus;
  createdDate: string;
  student?: User;
  bidCount: number;
}

export interface CreateQuestionRequest {
  title: string;
  description: string;
  subject: string;
  attachmentUrl?: string;
  preferredDuration: number;
  optionalBudget?: number;
}

export type UpdateQuestionRequestStatus =
  (typeof UpdateQuestionRequestStatus)[keyof typeof UpdateQuestionRequestStatus];

export const UpdateQuestionRequestStatus = {
  Open: "Open",
  Matched: "Matched",
  Scheduled: "Scheduled",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface UpdateQuestionRequest {
  status?: UpdateQuestionRequestStatus;
}

export type BidStatus = (typeof BidStatus)[keyof typeof BidStatus];

export const BidStatus = {
  Pending: "Pending",
  Accepted: "Accepted",
  Rejected: "Rejected",
} as const;

export interface Bid {
  bidId: number;
  questionId: number;
  tutorId: number;
  price: number;
  message: string;
  /** Duration in minutes */
  estimatedDuration: number;
  status: BidStatus;
  createdDate: string;
  tutor?: User;
  tutorProfile?: TutorProfile;
}

export interface CreateBidRequest {
  questionId: number;
  price: number;
  message: string;
  estimatedDuration: number;
}

export type UpdateBidRequestStatus =
  (typeof UpdateBidRequestStatus)[keyof typeof UpdateBidRequestStatus];

export const UpdateBidRequestStatus = {
  Pending: "Pending",
  Accepted: "Accepted",
  Rejected: "Rejected",
} as const;

export interface UpdateBidRequest {
  status: UpdateBidRequestStatus;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  Pending_Confirmation: "Pending Confirmation",
  Confirmed: "Confirmed",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface Session {
  sessionId: number;
  questionId: number;
  studentId: number;
  tutorId: number;
  proposedTime?: string | null;
  tutorCounterTime?: string | null;
  finalTime?: string | null;
  meetingLink?: string | null;
  status: SessionStatus;
  student?: User;
  tutor?: User;
  question?: Question;
}

export interface CreateSessionRequest {
  questionId: number;
  tutorId: number;
  proposedTime: string;
}

export type UpdateSessionRequestStatus =
  (typeof UpdateSessionRequestStatus)[keyof typeof UpdateSessionRequestStatus];

export const UpdateSessionRequestStatus = {
  Pending_Confirmation: "Pending Confirmation",
  Confirmed: "Confirmed",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface UpdateSessionRequest {
  proposedTime?: string;
  tutorCounterTime?: string;
  finalTime?: string;
  meetingLink?: string;
  status?: UpdateSessionRequestStatus;
}

export interface Review {
  reviewId: number;
  sessionId: number;
  studentId: number;
  tutorId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText?: string | null;
  createdDate: string;
  student?: User;
}

export interface CreateReviewRequest {
  sessionId: number;
  tutorId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText?: string;
}

export interface Notification {
  notificationId: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  read: boolean;
  relatedId?: number | null;
  createdDate: string;
}

export interface StudentDashboard {
  openQuestions: number;
  matchedQuestions: number;
  scheduledSessions: number;
  completedSessions: number;
  totalSpent: number;
  pendingBids: number;
  recentQuestions: Question[];
  upcomingSessions: Session[];
}

export interface SuspendUserRequest {
  suspended: boolean;
}

export interface TutorDashboard {
  openBids: number;
  acceptedBids: number;
  scheduledSessions: number;
  completedSessions: number;
  totalEarned: number;
  averageRating: number;
  recentBids: Bid[];
  upcomingSessions: Session[];
}

export type GetQuestionsParams = {
  status?: GetQuestionsStatus;
  subject?: string;
  studentId?: number;
};

export type GetQuestionsStatus =
  (typeof GetQuestionsStatus)[keyof typeof GetQuestionsStatus];

export const GetQuestionsStatus = {
  Open: "Open",
  Matched: "Matched",
  Scheduled: "Scheduled",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export type GetBidsParams = {
  questionId?: number;
  tutorId?: number;
  status?: GetBidsStatus;
};

export type GetBidsStatus = (typeof GetBidsStatus)[keyof typeof GetBidsStatus];

export const GetBidsStatus = {
  Pending: "Pending",
  Accepted: "Accepted",
  Rejected: "Rejected",
} as const;

export type GetSessionsParams = {
  studentId?: number;
  tutorId?: number;
  status?: GetSessionsStatus;
};

export type GetSessionsStatus =
  (typeof GetSessionsStatus)[keyof typeof GetSessionsStatus];

export const GetSessionsStatus = {
  Pending_Confirmation: "Pending Confirmation",
  Confirmed: "Confirmed",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export type GetReviewsParams = {
  tutorId: number;
};

export type AdminGetUsersParams = {
  role?: AdminGetUsersRole;
  suspended?: boolean;
};

export type AdminGetUsersRole =
  (typeof AdminGetUsersRole)[keyof typeof AdminGetUsersRole];

export const AdminGetUsersRole = {
  student: "student",
  tutor: "tutor",
  admin: "admin",
} as const;