# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TutorMatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /tutor-profiles/{tutorId}:
    get:
      operationId: getTutorProfile
      tags: [tutor-profiles]
      summary: Get tutor profile
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTutorProfile
      tags: [tutor-profiles]
      summary: Update tutor profile
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTutorProfileRequest"
      responses:
        "200":
          description: Tutor profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorProfile"
  /questions:
    get:
      operationId: getQuestions
      tags: [questions]
      summary: List questions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Open, Matched, Scheduled, Completed, Cancelled]
        - name: subject
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{questionId}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionRequest"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: List reviews for a tutor
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    put:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, tutor, admin]
        rating:
          type: number
          nullable: true
        suspended:
          type: boolean
        createdDate:
          type: string
          format: date-time
      required: [userId, name, email, role, suspended, createdDate]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
    TutorProfile:
      type: object
      properties:
        tutorId:
          type: integer
        subjects:
          type: array
          items:
            type: string
        education:
          type: string
          nullable: true
        experience:
          type: string
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/User"
      required: [tutorId, subjects]
    UpdateTutorProfileRequest:
      type: object
      properties:
        subjects:
          type: array
          items:
            type: string
        education:
          type: string
        experience:
          type: string
        hourlyRate:
          type: number
        bio:
          type: string
    Question:
      type: object
      properties:
        questionId:
          type: integer
        studentId:
          type: integer
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        preferredDuration:
          type: integer
          description: Duration in minutes
        optionalBudget:
          type: number
          nullable: true
        status:
          type: string
          enum: [Open, Matched, Scheduled, Completed, Cancelled]
        createdDate:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
        bidCount:
          type: integer
      required: [questionId, studentId, title, description, subject, preferredDuration, status, createdDate, bidCount]
    CreateQuestionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        attachmentUrl:
          type: string
        preferredDuration:
          type: integer
        optionalBudget:
          type: number
      required: [title, description, subject, preferredDuration]
    UpdateQuestionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Open, Matched, Scheduled, Completed, Cancelled]
...
    Review:
      type: object
      properties:
        reviewId:
          type: integer
        sessionId:
          type: integer
        studentId:
          type: integer
        tutorId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
          nullable: true
        createdDate:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
      required: [reviewId, sessionId, studentId, tutorId, rating, createdDate]
    CreateReviewRequest:
      type: object
      properties:
        sessionId:
          type: integer
        tutorId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
      required: [sessionId, tutorId, rating]
    Notification:
      type: object
      properties:
        notificationId:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        relatedId:
          type: integer
          nullable: true
        createdDate:
          type: string
          format: date-time
      required: [notificationId, userId, type, title, message, read, createdDate]
...