# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TutorMatch API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  sessionId: number;
  tutorId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText?: string;
}